graphql/requirements.txt
black
ecs-logging
fastapi
flake8
greenlet
oracledb
isort
mypy
pydantic
pydantic-settings
python-dotenv
pylance
pylint
sqlalchemy[asyncio]
strawberry-graphql[fastapi]
uvicorn
